The two branches each held half the product: the spine had the money loop
(Razorpay Test Mode execution, UNKNOWN as a settled-by-re-verification state,
reconciliation, the provider mapping table) and no identity control plane;
the trunk had RLS, OIDC, SCIM and relational roles and none of the money loop.
Neither was the product. This is the union.

Six defects the merge introduced or exposed, none of which a textual conflict
would have shown:

* Two alembic heads. c3f18a4d7b62 and 351d7cd7c50d were both written as
  children of d09395a87106, so `alembic upgrade head` refused to run. The
  spine's migration is re-pointed onto the identity chain.

* provider_mappings had no row-level policy. It carries a merchant_id and
  arrived from the branch that had no RLS, so it reached a tree that does
  with nothing filtering it -- and resolving another merchant's mapping is
  how a refund lands on somebody else's money. Added to MERCHANT_SCOPED with
  its own migration. `test_every_table_with_a_merchant_is_covered` is what
  caught it, which is the test doing exactly its job.

* Two `class Readiness`. Both branches defined a response model by that name
  with different shapes; the later definition won and /readiness returned 500
  on response validation. The trunk's is now InstanceReadiness.

* The seeder stopped inserting provider_mappings. Resolving toward the
  trunk's roles-aware `insert_all` was right, but its INSERT_ORDER had never
  heard of the table. Both branches had already been bitten by a copy of this
  list drifting; there is now one list.

* Timeline was the last file on React Router v6. package.json did not
  conflict, so the merge took the spine's v7 and dropped react-router-dom
  while keeping the trunk's file that imported it.

* CI installed from requirements.lock, which resolves without redis. The two
  branches had opposite meanings for requirements.txt -- the trunk's is the
  hashed output, the spine's was a `>=` declaration. The trunk's scheme wins
  because the spine's lock cannot install the trunk's stack; the orphaned
  lock files are dropped rather than left as a trap.

Measured on the merged tree: 1072 passed / 14 skipped, 324 Vitest, 187/187
scenarios with 130/130 critical. The mutation corpus is the union at 136 and
is published as defined rather than caught -- no complete run has measured
this tree, and CI's nightly is where that figure will come from.

+++ app/recovery/ledger.py
@@ -129,12 +129,23 @@ def build_ledger(session, merchant_id: str) -> RecoveryLedger:
     led.outstanding_minor = int(row["outstanding"])
 
     # §50: at risk BY INCIDENT and BY PAYMENT METHOD.
+    #
+    # `::bigint` on every SUM, and it is not decoration. Postgres widens
+    # `SUM()` over a bigint to `numeric`, psycopg2 renders that as `Decimal`,
+    # and these two breakdowns were the only place in the ledger where money
+    # reached the wire without a declared integer field to be coerced against.
+    # The result was `recoverable_minor` arriving as the STRING "2798847" here
+    # while the identically-named field one level up was an integer -- money on
+    # a revenue ledger with two types in one response. The response models now
+    # declare `int` as well; this makes the value an integer before pydantic
+    # ever sees it, so the wire type stops depending on what Postgres decides a
+    # sum should widen to.
     led.by_incident = [dict(r) for r in session.execute(text("""
         SELECT i.id AS incident_id, i.incident_type, i.severity, i.status, i.title,
                i.revenue_at_risk_minor,
                COALESCE(SUM(c.attributed_amount_minor) FILTER (
-                   WHERE c.status <> 'INELIGIBLE'), 0) AS recoverable_minor,
-               COALESCE(SUM(c.actual_recovery_minor), 0) AS recovered_minor
+                   WHERE c.status <> 'INELIGIBLE'), 0)::bigint AS recoverable_minor,
+               COALESCE(SUM(c.actual_recovery_minor), 0)::bigint AS recovered_minor
         FROM incidents i LEFT JOIN recovery_candidates c ON c.incident_id = i.id
         WHERE i.merchant_id = :m AND i.status = ANY(:open)
         GROUP BY i.id, i.incident_type, i.severity, i.status, i.title,
@@ -147,8 +158,8 @@ def build_ledger(session, merchant_id: str) -> RecoveryLedger:
     led.by_method = [dict(r) for r in session.execute(text("""
         SELECT COALESCE(p.method, 'unknown') AS method,
                COALESCE(SUM(c.attributed_amount_minor) FILTER (
-                   WHERE c.status <> 'INELIGIBLE'), 0) AS recoverable_minor,
-               COALESCE(SUM(c.actual_recovery_minor), 0) AS recovered_minor,
+                   WHERE c.status <> 'INELIGIBLE'), 0)::bigint AS recoverable_minor,
+               COALESCE(SUM(c.actual_recovery_minor), 0)::bigint AS recovered_minor,
                COUNT(*) AS candidates
         FROM recovery_candidates c JOIN payments p ON p.id = c.payment_id
         WHERE c.merchant_id = :m
